requirements.txt
flask
flask-cors
openai
pandas
python-dotenv
requests
beautifulsoup4
openpyxl
numpy
scikit-learn
Pillow